Pass current directory as paths to pep8 StyleGuide
Instead of setting parse_argv to True, which fails if no config files
are present, we pass the current directory as the paths keyword
argument when creating the pep8 StyleGuide. This causes pep8 to parse
local config files as expected and works also if none are present.

Requires pep8 68d72a5d or newer.