Make --install-hook work without config files
In short, pep8's option processing would bomb out if there were neither args
provided or config files present (setup.cfg, tox.ini, .pep8, ...).
--install-hook should work regardless of whether a config file is present or
not, so we use a callback to add an argument to the parser's internal argument
list and skip the check altogether. Links for context are in the comments in
this diff.