Fix hook test on Windows
This test was failing on Windows, for several reasons. Number one, the paths
for the hook test were hard-coded with forward slashes as path separators. I
replaced those with calls to os.path.join. Number two, the drive letter was
getting in the way of the path manipulation. On Windows, I'm using
os.path.splitdrive to fix that.