Fix parsing ignore
https://gitlab.com/pycqa/flake8/issues/40

* Adding an ignore option in [flake8] wasn't working because pep8.StyleGuide
turned the string sent in into a tuple, which the option parser needs to receive
as an iterable that isn't a string. Split on spaces, commas, or semicolons using
re.findall in order to get a list of error/warnings to pass to StyleGuide
properly.

* Add self to contributors

(cherry picked from commit 8f8a7d458f8f46669398591bffd93462d4e439ed)