Handle repeated --quiet options again
Fundamentally on Flake8 2.x using -q altered the format of the errors
(and the behaviour a little) so it makes the most sense to implement
this logic with formatters rather than messy logic spread throughout
the project.

The FilenameOnly formatter will keep track of filenames already reported
and only print the name once while Nothing will print nothing.

Closes #180