Handle AttributeErrors during parameter aggregation
Plugins do not have their parameters checked in advance, so when we
try to aggregate parameters for a plugin, there's a chance it may
request an attribute of the FileProcessor that simply does not exist.

We catch this and re-raise it but we should also capture it in the
checker manager and handle it appropriately there as well.