Remove CONTRIBUTING.rst symlink
The GitLab interface links to the CONTRIBUTING.rst file
(e.g. next to the green submit button under the text box
for filing a new issue), but does not follow the symlink.

Removing the symlink in order to replace it with a file
pointing at the rendered version of the guidelines:
http://flake8.pycqa.org/en/latest/internal/contributing.html

See GitLab issue #327.