Use extras_require instead of conditional code.
Use of extras_require produces correct package metadata for flake8
regardless of the version of Python the package is built with.
The conditional code produces different metadata for Python 3 and 2.
The latest versions of flake8 were released with Python 3 and metadata
did not contain two dependencies necessary for use with Python 2.7.

See https://gitlab.com/pycqa/flake8/issues/341 for details.

Account for users of setuptools<18 that still need the conditional code.