FIX: Change how partial-path file patterns match
Previously, filenames with path separators in them will be made
absolute before matching.  However, this causes files specified in a
configuration file to match different files based on where in the code
base you are running flake8 from.

This changes to pre-pending a '*' to anything with a path in it.  This
restores the behavior of per-file-ignores plugin matching.