Update information on default ignore list
The paragraph directly under "Ignoring Violations with Flake8" was a
less specific version of the same information under the last paragraph
of "Extending the Default Ignore List." To remove the duplication, I
suggest combining the two paragraphs together and giving it its own
heading to make sure it stands out. But, if there's a better way of
organizing, I'm open to suggestions and can revise this accordingly.

Add a link to the glossary for error classes.