README - add a flake8 project badge
Add a project badge image and display on the README.
This gives people a way to show they use Flake8
as the linter for their project,
and link to the Flake8 project page.

This may be a fun way for people to show their support,
and help drive new users to Flake8.